requirements.txt
# Core frameworks
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart
websockets

# AI and ML
openai>=1.50.0
pathway[all]>=0.13.0
llama-index-embeddings-openai>=0.3.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0

# LangChain DeepAgents
langgraph>=0.2.0

# MongoDB
motor>=3.6.0
pymongo>=4.10.0

# Vector Database
pinecone-client>=6.0.0

# LandingAI (required for document parsing)
landingai-ade>=0.15.0

# Document processing
pypdf>=5.0.0
python-docx>=1.1.0
openpyxl>=3.1.0
pillow>=10.0.0
unstructured>=0.15.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
httpx>=0.27.0
nest-asyncio>=1.5.0
